refactor: [윤영현] request, resposne 패키지 service 안으로 이동 [HEALTH-FOOD-PROCESS]
