+++ src/main/java/com/example/sbsj_process/account/controller/form/MemberLoginForm.java
@@ -1,6 +1,6 @@
 package com.example.sbsj_process.account.controller.form;
 
-import com.example.sbsj_process.account.request.MemberLoginRequest;
+import com.example.sbsj_process.account.service.request.MemberLoginRequest;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
