+++ src/main/java/com/example/sbsj_process/account/controller/form/MemberRegisterForm.java
@@ -1,6 +1,6 @@
 package com.example.sbsj_process.account.controller.form;
 
-import com.example.sbsj_process.account.request.MemberRegisterRequest;
+import com.example.sbsj_process.account.service.request.MemberRegisterRequest;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
