+++ src/main/java/com/example/sbsj_process/account/service/MemberService.java
@@ -1,12 +1,12 @@
 package com.example.sbsj_process.account.service;
 
 
-import com.example.sbsj_process.account.request.MemberCheckPasswordRequest;
-import com.example.sbsj_process.account.request.MemberLoginRequest;
-import com.example.sbsj_process.account.request.MemberRegisterRequest;
-import com.example.sbsj_process.account.request.MyPageUpdateRequest;
-import com.example.sbsj_process.account.response.MemberInfoResponse;
-import com.example.sbsj_process.account.response.MemberLoginResponse;
+import com.example.sbsj_process.account.service.request.MemberCheckPasswordRequest;
+import com.example.sbsj_process.account.service.request.MemberLoginRequest;
+import com.example.sbsj_process.account.service.request.MemberRegisterRequest;
+import com.example.sbsj_process.account.service.request.MyPageUpdateRequest;
+import com.example.sbsj_process.account.service.response.MemberInfoResponse;
+import com.example.sbsj_process.account.service.response.MemberLoginResponse;
 
 public interface MemberService {
