+++ src/main/java/com/example/sbsj_process/account/service/MemberServiceImpl.java
@@ -7,12 +7,12 @@ import com.example.sbsj_process.account.entity.MemberProfile;
 import com.example.sbsj_process.account.repository.AuthenticationRepository;
 import com.example.sbsj_process.account.repository.MemberProfileRepository;
 import com.example.sbsj_process.account.repository.MemberRepository;
-import com.example.sbsj_process.account.request.MemberCheckPasswordRequest;
-import com.example.sbsj_process.account.request.MemberLoginRequest;
-import com.example.sbsj_process.account.request.MemberRegisterRequest;
-import com.example.sbsj_process.account.request.MyPageUpdateRequest;
-import com.example.sbsj_process.account.response.MemberInfoResponse;
-import com.example.sbsj_process.account.response.MemberLoginResponse;
+import com.example.sbsj_process.account.service.request.MemberCheckPasswordRequest;
+import com.example.sbsj_process.account.service.request.MemberLoginRequest;
+import com.example.sbsj_process.account.service.request.MemberRegisterRequest;
+import com.example.sbsj_process.account.service.request.MyPageUpdateRequest;
+import com.example.sbsj_process.account.service.response.MemberInfoResponse;
+import com.example.sbsj_process.account.service.response.MemberLoginResponse;
 import com.example.sbsj_process.order.repository.DeliveryRepository;
 import com.example.sbsj_process.security.service.RedisService;
 import lombok.RequiredArgsConstructor;
