+++ src/main/java/com/example/sbsj_process/account/service/request/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.example.sbsj_process.account.request;
+package com.example.sbsj_process.account.service.request;
 
 import com.example.sbsj_process.account.entity.Member;
 import com.example.sbsj_process.account.entity.MemberProfile;
