+++ src/main/java/com/example/sbsj_process/order/controller/DeliveryController.java
@@ -1,7 +1,7 @@
 package com.example.sbsj_process.order.controller;
 
-import com.example.sbsj_process.order.request.DeliveryRegisterRequest;
-import com.example.sbsj_process.order.response.DeliveryListResponse;
+import com.example.sbsj_process.order.service.request.DeliveryRegisterRequest;
+import com.example.sbsj_process.order.service.response.DeliveryListResponse;
 import com.example.sbsj_process.order.service.DeliveryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
