+++ src/main/java/com/example/sbsj_process/order/service/DeliveryService.java
@@ -1,7 +1,7 @@
 package com.example.sbsj_process.order.service;
 
-import com.example.sbsj_process.order.request.DeliveryRegisterRequest;
-import com.example.sbsj_process.order.response.DeliveryListResponse;
+import com.example.sbsj_process.order.service.request.DeliveryRegisterRequest;
+import com.example.sbsj_process.order.service.response.DeliveryListResponse;
 
 import java.util.List;
