+++ src/main/java/com/example/sbsj_process/order/service/DeliveryServiceImpl.java
@@ -5,8 +5,8 @@ import com.example.sbsj_process.account.repository.MemberProfileRepository;
 import com.example.sbsj_process.account.repository.MemberRepository;
 import com.example.sbsj_process.order.entity.Delivery;
 import com.example.sbsj_process.order.repository.DeliveryRepository;
-import com.example.sbsj_process.order.request.DeliveryRegisterRequest;
-import com.example.sbsj_process.order.response.DeliveryListResponse;
+import com.example.sbsj_process.order.service.request.DeliveryRegisterRequest;
+import com.example.sbsj_process.order.service.response.DeliveryListResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
