+++ src/main/java/com/example/sbsj_process/product/service/ProductService.java
@@ -1,7 +1,7 @@
 package com.example.sbsj_process.product.service;
 
 import com.example.sbsj_process.product.controller.form.ProductListResponse;
-import com.example.sbsj_process.product.request.ProductRegisterRequest;
+import com.example.sbsj_process.product.service.request.ProductRegisterRequest;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
