+++ src/main/java/com/example/sbsj_process/product/service/request/ProductRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.example.sbsj_process.product.request;
+package com.example.sbsj_process.product.service.request;
 
 import com.example.sbsj_process.product.entity.Product;
 import com.example.sbsj_process.product.entity.ProductInfo;
