+++ src/test/java/com/example/sbsj_process/AccountTest/memberTest.java
@@ -2,9 +2,9 @@ package com.example.sbsj_process.AccountTest;
 
 import com.example.sbsj_process.account.repository.AuthenticationRepository;
 import com.example.sbsj_process.account.repository.MemberRepository;
-import com.example.sbsj_process.account.request.MemberLoginRequest;
-import com.example.sbsj_process.account.request.MemberRegisterRequest;
-import com.example.sbsj_process.account.response.MemberLoginResponse;
+import com.example.sbsj_process.account.service.request.MemberLoginRequest;
+import com.example.sbsj_process.account.service.request.MemberRegisterRequest;
+import com.example.sbsj_process.account.service.response.MemberLoginResponse;
 import com.example.sbsj_process.account.service.MemberService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
